+++ src/main/java/me/dantaeusb/zetter/entity/item/PaintingEntity.java
@@ -346,7 +346,7 @@ public class PaintingEntity extends HangingEntity implements IEntityAdditionalSp
      * Copied from PaintingEntity
      */
     public void lerpTo(double x, double y, double z, float yaw, float pitch, int posRotationIncrements, boolean teleport) {
-        BlockPos blockpos = this.pos.offset(x - this.getX(), y - this.getY(), z - this.getZ());
+        BlockPos blockpos = this.pos.offset((int) (x - this.getX()), (int) (y - this.getY()), (int) (z - this.getZ()));
         this.setPos(blockpos.getX(), blockpos.getY(), blockpos.getZ());
     }
